Get newer Cabal for Hadrian, use for `findPtr`
There are many more things we would use this for, this is just a first
simple example to ensure the Cabal bump is working as intended.

The support for multiple package in `bootstrap.py` matches Cabal's
`bootstrap.py`, its inspiration, in order to keep it easy to sync
changes back and forth.